Start splitting the action source
