+++ os/server/ktor/src/jvmMain/kotlin/com/wasmo/ktor/HttpActionSource.kt
@@ -1,6 +1,10 @@
 package com.wasmo.ktor
 
+import com.wasmo.framework.HttpActionBinder
+
 interface HttpActionSource {
+  val order: Int
+
   context(binder: HttpActionBinder)
   fun bindActions()
 }
